refactor(docs): extract content-block snippets into lib/snippets.ts
Pure relocation of the in-place `<!-- docs:* -->` marker rendering out of
scripts/docs/generate.ts: commandSnippets + docsSnippetBlocks consts, the five
snippet renderers (renderCommandSnippet / renderApiRoutesSnippet /
renderSettingsSnippet / renderPluginPermissionsSnippet / renderDocsSnippetBlock),
and the page-rewriter updateGeneratedContentBlocks. docsRoot is recomputed
locally in the module. Function bodies byte-for-byte unchanged; generate.ts
imports updateGeneratedContentBlocks back (the renderers are module-internal).
Dropped the now-unused node:fs, getApiRoutes/getCliCommands/extractPluginSettings/
renderExecToolsSnippet, permissions, and escapeTableCell imports.

Verified: docs:generate reference output byte-identical AND the full docs diff
(including the 10 marker-block pages updateGeneratedContentBlocks rewrites) is
identical to the parent branch.
